powerctl: Re-locate high-voltage threshold check.
If the charger is doing a bulk charge, it'll likely be dumping high
current into the battery, resulting in a high voltage appearing at the
terminals.

We don't want to disturb this… we want to wait until the charger thinks
it's done, which it signifies by turning off.  We should see the voltage
drop back and continue falling.  If we're above our "high" threshold at
this point, *then* we might consider leaving it off.