setpoints: Add warning timeout.
If the charge isn't rising, we should wait before acting on it in case
it is temporary in nature.

It was hoped that the charger delay should be enough, but it seems it
still can flap.  We reset the warning if we see an improvement, or let
the timer expire if we don't.