setpoints: Drop solar to 4V, battery to 12.75V
Reduce the solar threshold so that we're at the point where the SSR can
work properly.  i.e. we're at first light.

The battery can discharge for an hour or two at this point, which will
give it a break from the charger just before sunrise.  This may trigger
the other charge controller to charge the battery during the day.

Drop the battery voltage just a fraction so we are past the "knee" of
the discharge curve at 12.8V.